Drop `go-memdb` and use a hand-written optimized version for our particular usecase.

Breaking change: replace `*string` as controller.Input.ID with `optional.Optional` for readability, comparison and ease of use.

Fixes #305