@smira 2023-09-22
Rough comparison:

```
name                        old time/op    new time/op    delta
GetDependentControllers-32    1.11µs ± 2%    0.09µs ±10%  -91.57%  (p=0.029 n=4+4)
BuildDatabase-32              73.0µs ± 5%    90.9µs ± 4%  +24.52%  (p=0.029 n=4+4)

name                        old alloc/op   new alloc/op   delta
GetDependentControllers-32      464B ± 0%       32B ± 0%  -93.10%  (p=0.029 n=4+4)
BuildDatabase-32              48.3kB ± 1%     1.2kB ± 3%  -97.43%  (p=0.029 n=4+4)

name                        old allocs/op  new allocs/op  delta
GetDependentControllers-32      20.0 ± 0%       1.0 ± 0%  -95.00%  (p=0.029 n=4+4)
BuildDatabase-32                 850 ± 1%         7 ± 0%  -99.18%  (p=0.029 n=4+4)
```