@smira 2023-09-25
Updated benchmark (benchmark itself slightly changed as well):

```
name                        old time/op    new time/op    delta
GetDependentControllers-32    1.08µs ± 4%    0.11µs ± 5%  -90.08%  (p=0.029 n=4+4)
BuildDatabase-32              71.8µs ± 6%     6.1µs ± 2%  -91.45%  (p=0.029 n=4+4)

name                        old alloc/op   new alloc/op   delta
GetDependentControllers-32      464B ± 0%       32B ± 0%  -93.10%  (p=0.029 n=4+4)
BuildDatabase-32              47.9kB ± 1%     1.3kB ± 8%  -97.30%  (p=0.029 n=4+4)

name                        old allocs/op  new allocs/op  delta
GetDependentControllers-32      20.0 ± 0%       1.0 ± 0%  -95.00%  (p=0.029 n=4+4)
BuildDatabase-32                 843 ± 1%         8 ± 0%  -99.05%  (p=0.029 n=4+4)
```